feat: [DPAV-481] debrand application
