+++ address-profiling-pipeline/initial_load_adapter/infrastructure/Dockerfile
