+++ address-profiling-pipeline/initial_load_adapter/infrastructure/job.yaml
