+++ address-profiling-pipeline/initial_load_adapter/src/producer.py
