+++ address-profiling-pipeline/mapper/src/address_profile_to_knowledge_mapper.py
@@ -24,39 +24,24 @@ from json import loads
 from mapping_function import map_func
 from dotenv import load_dotenv
 
+# run adapter.py, then mapper.py, then mapper-2.py and finally, this file
+
+# Mapper Configuration
 load_dotenv()
 config = Configurator()
-BROKER = config.get(
-    "BOOTSTRAP_SERVERS",
-    required=True,
-    description="Specifies the Kafka Bootstrap Servers to connect to.",
-)
-SASL_USERNAME = config.get(
-    "SASL_USERNAME", required=True,
-    description="The username for the SASL authentication."
-)
-SASL_PASSWORD = config.get(
-    "SASL_PASSWORD", required=True,
-    description="The password for the SASL authentication."
-)
-SOURCE_TOPIC = config.get(
-    "SOURCE_TOPIC",
-    required=True,
-    description="Specifies the Kafka topic the mapper ingests from.",
-)
-SOURCE_TOPIC_GROUP_ID = config.get(
-    "SOURCE_TOPIC_GROUP_ID",
-    required=True,
-    description="Specifies the Kafka topic group id the mapper ingests from.",
-)
-TARGET_TOPIC = config.get(
-    "TARGET_TOPIC",
-    required=True,
-    description="Specifies the Kafka topic the mapper pushes its output to",
-)
-DEBUG = config.get(
-    "DEBUG", required=False, default=False, converter=bool, required_type=bool
-)
+BROKER = config.get("BOOTSTRAP_SERVERS", required=True,
+                    description="Specifies the Kafka Bootstrap Servers to connect to.")
+SASL_USERNAME = config.get("SASL_USERNAME", required=True,
+                    description="The username for the SASL authentication.")
+SASL_PASSWORD = config.get("SASL_PASSWORD", required=True,
+                    description="The password for the SASL authentication.")
+SOURCE_TOPIC = config.get("SOURCE_TOPIC", required=True,
+                    description="Specifies the Kafka topic the mapper ingests from.")
+TARGET_TOPIC = config.get("TARGET_TOPIC", required=True,
+                    description="Specifies the Kafka topic the mapper pushes its output to")
+SOURCE_TOPIC_GROUP_ID = config.get("SOURCE_TOPIC_GROUP_ID", required=False,
+                    description="The group id for the topic to consume", default=0)
+DEBUG = config.get("DEBUG", required=False, default=False, converter=bool, required_type=bool)
 
 kafka_config = {
     "bootstrap.servers": BROKER,
@@ -79,47 +64,22 @@ kafka_producer_config = {
 logger = CoreLoggerFactory.get_logger(
     "{source}-to-{target}-mapper".format(source=SOURCE_TOPIC, target=TARGET_TOPIC),
     kafka_config=kafka_producer_config,
-    topic="telicent-logging",
+    topic="logging",
 )
-
-
 # Function each record on the source topic is passed to.
-def mapping_function(record: Record) -> Union[Record, List[Record], None]:
+def mapping_function(record: Record) ->  Union[Record, List[Record], None]:
+    
+
     # Add mapping logic here
     # Try to keep it small and performant
     data = loads(record.value)
-
+    
     mapped = map_func(data)
-    if mapped is None:
-        logger.warning(
-            "{uprn}, {address} does not container a lat/lon lookup".format(
-                uprn=data["UPRN"], address=data["Address"]
-            ),
-        )
-        print(
-            "{uprn}, {address} does not container a lat/lon lookup".format(
-                uprn=data["UPRN"], address=data["Address"]
-            )
-        )
-        return mapped
-    return RecordUtils.add_header(
-        Record(record.headers, record.key, mapped, None),
-        "Content-Type",
-        "text/turtle",
-    )
+
+    return RecordUtils.add_header(Record(record.headers, record.key, mapped, None), "Content-Type", "application/n-triples")
 
 
-source = KafkaSource(
-    topic=SOURCE_TOPIC, kafka_config=kafka_config
-)
+source = KafkaSource(topic=SOURCE_TOPIC, kafka_config=kafka_config)
 target = KafkaSink(topic=TARGET_TOPIC, kafka_config=kafka_producer_config)
-mapper = Mapper(
-    source,
-    target,
-    mapping_function,
-    name=SOURCE_TOPIC + " to " + TARGET_TOPIC + " Mapper",
-    has_reporter=False,
-    reporting_batch_size=500,
-    has_error_handler=False
-)
+mapper = Mapper(source, target, mapping_function, name=SOURCE_TOPIC + " to " + TARGET_TOPIC + " Mapper", has_reporter=False, reporting_batch_size=500, has_error_handler=False)
 mapper.run()
