+++ address-profiling-pipeline/single_file_s3_producer/infrastructure/Dockerfile
