+++ lat-long-pipeline/adapter/infrastructure/Dockerfile
