+++ lat-long-pipeline/adapter/infrastructure/job.yaml
