+++ lat-long-pipeline/mapper/infrastructure/Dockerfile
